noon push

+++ MapComps/PlacePreview.js
@@ -2,8 +2,8 @@
 
 import MapService from '../services/MapService.js';
 
-{/* <button class="delete-btn" @click="deletePlace(place.id)">X</button> */}
-{/* <button class="render-details" @click="$emit('editClicked', place.id)">See More</button> */}
+{/* <button class="delete-btn" @click="deletePlace(place.id)">X</button> */ }
+{/* <button class="render-details" @click="$emit('editClicked', place.id)">See More</button> */ }
 
 export default {
     template: `
@@ -12,47 +12,47 @@ export default {
         <h4>{{place.tag}}</h4>
         <div class="action-btns">
             <i class="fa fa-expand render-details" aria-hidden="true" @click="$emit('editClicked', place.id)"></i> 
-            <i class="fa fa-map-marker" aria-hidden="true"></i>            
+            <i class="fa fa-map-marker" aria-hidden="true" @click="showSelectedPlace"></i>            
             <i class="fa fa-trash-o delete-btn" aria-hidden="true" @click="deletePlace(place.id)"></i>
         </div> 
         </section>
     `,
-    props: ['place'], 
+    props: ['place'],
     data() {
         return {
-            updatedPlace: {}, 
+            updatedPlace: {},
         }
     },
     methods: {
         deletePlace(placeId) {
-            MapService.deletePlace(placeId)
-            .then(_ => {
-                alert('Are you sure about this?');
-
-                // SweetAlert function
-                // swal({
-                //     title: 'Are you sure?',
-                //     text: "You won't be able to revert this!",
-                //     type: 'warning',
-                //     showCancelButton: true,
-                //     confirmButtonColor: '#3085d6',
-                //     cancelButtonColor: '#d33',
-                //     confirmButtonText: 'Yes, delete it!'
-                //   }).then((result) => {
-                //     if (result.value) {
-                //       swal(
-                //         'Deleted!',
-                //         'Your file has been deleted.',
-                //         'success'
-                //       )
-                //     }
-                //   })
+            // Assure delete with SweetAlert
+            swal({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                type: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+            }).then((result) => {
+                if (result.value) {
+                    MapService.deletePlace(placeId)
+                    swal(
+                        'Deleted!',
+                        'Your file has been deleted.',
+                        'success'
+                    )
+                }
             })
         },
+        showSelectedPlace() {
+            MapService.initMap(this.place.lat, this.place.lng);
+        }
         // showPlaceOnMap() {
         //     MapService.initMap(res.geometry.location.lat, res.geometry.location.lng);
         // }
-    }
+    },
+
 }
